[Fix] Unify dataclasses under a types.py for ease
