+++ tests/chunker/test_semantic_chunker.py
@@ -3,9 +3,9 @@ from typing import List
 
 import pytest
 
-from chonkie import Chunk, SemanticChunk, SemanticChunker
+from chonkie import SemanticChunker
 from chonkie.embeddings import Model2VecEmbeddings, OpenAIEmbeddings
-
+from chonkie.types import Chunk, SemanticChunk
 
 @pytest.fixture
 def sample_text():
